Rename and move run_command() to utils module
