+++ scripts/gcp_setup/utils.py
@@ -1,0 +1,40 @@
+# Copyright 2017 The Forseti Security Authors. All rights reserved.
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
+"""Utils.
+
+This has been tested with python 2.7.
+"""
+
+from __future__ import print_function
+
+import subprocess
+
+
+def run_command(cmd_args):
+    """Wrapper to run a command in subprocess.
+
+    Args:
+        cmd_args (str): The list of command arguments.
+
+    Returns:
+        int: The return code. 0 is "ok", anything else is "error".
+        str: Output, if command was successful.
+        err: Error output, if there was an error.
+    """
+    proc = subprocess.Popen(cmd_args,
+                            stdout=subprocess.PIPE,
+                            stderr=subprocess.PIPE)
+    out, err = proc.communicate()
+    return proc.returncode, out, err
